requirements.txt
Flask
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-core
langchain-text-splitters
pillow
facebook-sdk
requests